% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{cbo2002_isco88}
\alias{cbo2002_isco88}
\title{Correspondência da CBO-2002 com a ISCO-88, por ocupação}
\format{
\code{data.frame} com as colunas:
\describe{
\item{cbo2002}{código de seis dígitos, sem hífen, como nos microdados.}
\item{titulo}{título da ocupação na CBO-2002.}
\item{familia}{os quatro primeiros dígitos.}
\item{cbo94}{código correspondente na CBO-94, sem pontuação.}
\item{isco88}{código ISCO-88 de quatro dígitos.}
}
}
\source{
Tábua de conversão CBO2002--CBO94--CIUO88 do Ministério do Trabalho,
consultada família a família em
\url{http://www.mtecbo.gov.br/cbosite/pages/tabua/FiltroConversao_CBO2002_CBO94_CIUO88.jsf}
}
\usage{
cbo2002_isco88
}
\description{
Correspondência da CBO-2002 com a ISCO-88, por ocupação
}
\section{Cobertura}{

A tábua do Ministério do Trabalho é uma conversão entre a CBO-2002 e a
CBO-94, de modo que cobre apenas as ocupações presentes nas duas
classificações. As ocupações criadas na revisão de 2002 e o grande grupo 0
(forças armadas) não têm correspondência oficial com a CIUO-88 e não
aparecem aqui.
}

\keyword{datasets}
