% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbo.R
\name{cbo2002_para_isei}
\alias{cbo2002_para_isei}
\title{Indice socioeconomico ISEI a partir da CBO-2002}
\usage{
cbo2002_para_isei(cbo)
}
\arguments{
\item{cbo}{Vetor de codigos da CBO-2002, com 6 digitos (ocupacao) ou 4
(familia). Aceita "1111-05", "111105" ou 111105.}
}
\value{
Vetor numerico com o escore ISEI-88.
}
\description{
Indice socioeconomico ISEI a partir da CBO-2002
}
\examples{
cbo2002_para_isei(c("1111-05", "225120"))
}
