% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbo.R
\name{cbo94_para_isco}
\alias{cbo94_para_isco}
\title{Traduz a CBO-94 em ISCO-88}
\usage{
cbo94_para_isco(cbo94)
}
\arguments{
\item{cbo94}{Vetor de codigos da CBO-94. Aceita "2-11.20" ou "21120".}
}
\value{
Vetor de texto com o ISCO-88 de quatro digitos, ou \code{NA}.
}
\description{
Para microdado anterior a 2003, que usa a classificacao de 1994 (cinco
digitos). A correspondencia vem da mesma tabua oficial do Ministerio do
Trabalho, pela coluna CBO-94.
}
\examples{
cbo94_para_isco(c("2-11.20", "21130"))
}
