% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbo.R
\name{cbo94_para_isei}
\alias{cbo94_para_isei}
\title{Indice socioeconomico ISEI a partir da CBO-94}
\usage{
cbo94_para_isei(cbo94)
}
\arguments{
\item{cbo94}{Vetor de codigos da CBO-94. Aceita "2-11.20" ou "21120".}
}
\value{
Vetor numerico com o escore ISEI-88.
}
\description{
Indice socioeconomico ISEI a partir da CBO-94
}
\examples{
cbo94_para_isei("2-11.20")
}
