% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isei_br.R
\name{cbo94_para_isei_br}
\alias{cbo94_para_isei_br}
\title{ISEI-BR a partir da CBO-94}
\usage{
cbo94_para_isei_br(cbo94)
}
\arguments{
\item{cbo94}{Vetor de codigos da CBO-94. Aceita "2-11.20" ou "21120".}
}
\value{
Vetor numerico com o escore ISEI-BR, entre 10 e 90.
}
\description{
Versao para a RAIS anterior a 2003. Leia \code{\link[=tse_para_isei_br]{tse_para_isei_br()}} para o que a
regua e e o que ela nao e.
}
\examples{
cbo94_para_isei_br("2-11.20")
}
\seealso{
\code{\link[=tse_para_isei_br]{tse_para_isei_br()}}, \link{isco08_isei_br}
}
