% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prestigio.R
\name{cbo94_para_prestigio}
\alias{cbo94_para_prestigio}
\title{Prestigio de Treiman a partir da CBO-94}
\usage{
cbo94_para_prestigio(cbo94)
}
\arguments{
\item{cbo94}{Vetor de codigos da CBO-94. Aceita "2-11.20" ou "21120".}
}
\value{
Vetor numerico.
}
\description{
Prestigio de Treiman a partir da CBO-94
}
\examples{
cbo94_para_prestigio("2-11.20")
}
