% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbo.R
\name{cbo94_para_siops}
\alias{cbo94_para_siops}
\title{Prestigio ocupacional de Treiman a partir da CBO-94}
\usage{
cbo94_para_siops(cbo94)
}
\arguments{
\item{cbo94}{Vetor de codigos da CBO-94. Aceita "2-11.20" ou "21120".}
}
\value{
Vetor numerico com o escore de prestigio.
}
\description{
Prestigio ocupacional de Treiman a partir da CBO-94
}
\examples{
cbo94_para_siops("2-11.20")
}
