% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbo.R
\name{checa_cobertura_cbo94}
\alias{checa_cobertura_cbo94}
\title{Verifica a cobertura de codigos da CBO-94}
\usage{
checa_cobertura_cbo94(cbo94, silencioso = FALSE)
}
\arguments{
\item{cbo94}{Vetor de codigos da CBO-94 observados no seu dado.}

\item{silencioso}{Se \code{TRUE}, nao escreve a mensagem de sucesso.}
}
\value{
Invisivelmente, \code{TRUE}.
}
\description{
Equivalente de \code{\link[=checa_cobertura_cbo2002]{checa_cobertura_cbo2002()}} para o microdado anterior a 2003.
}
\examples{
checa_cobertura_cbo94(c("2-11.20", "21130"))
}
