% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod.R
\name{checa_cobertura_cod}
\alias{checa_cobertura_cod}
\title{Verifica a cobertura de codigos da COD}
\usage{
checa_cobertura_cod(cod, silencioso = FALSE)
}
\arguments{
\item{cod}{Vetor de codigos da COD observados no seu dado.}

\item{silencioso}{Se \code{TRUE}, nao escreve a mensagem de sucesso.}
}
\value{
Invisivelmente, \code{TRUE}.
}
\description{
Verifica a cobertura de codigos da COD
}
\examples{
checa_cobertura_cod(c("2211", "0411"))
}
