% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod.R
\name{cod_para_isei}
\alias{cod_para_isei}
\title{ISEI-88 a partir da COD}
\usage{
cod_para_isei(cod)
}
\arguments{
\item{cod}{Vetor de codigos da COD, de quatro digitos.}
}
\value{
Vetor numerico com o escore ISEI-88.
}
\description{
Use este, e nao \code{\link[=cod_para_isei08]{cod_para_isei08()}}, para comparar com dado eleitoral: o
pacote e ancorado na ISCO-88 e e nela que o TSE aterrissa.
}
\examples{
cod_para_isei(c("2211", "9629"))
}
