% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod.R
\name{cod_para_isei08}
\alias{cod_para_isei08}
\title{ISEI-08 a partir da COD}
\usage{
cod_para_isei08(cod)
}
\arguments{
\item{cod}{Vetor de codigos da COD, de quatro digitos.}
}
\value{
Vetor numerico com o escore ISEI-08.
}
\description{
ISEI-08 a partir da COD
}
\examples{
cod_para_isei08(c("2211", "9629"))
}
