% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod.R
\name{cod_para_prestigio08}
\alias{cod_para_prestigio08}
\title{Prestigio de Treiman ancorado na ISCO-08, a partir da COD}
\usage{
cod_para_prestigio08(cod)
}
\arguments{
\item{cod}{Vetor de codigos da COD, de quatro digitos.}
}
\value{
Vetor numerico.
}
\description{
Prestigio de Treiman ancorado na ISCO-08, a partir da COD
}
\examples{
cod_para_prestigio08(c("2211", "9629"))
}
