% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbo.R
\name{isco88_para_isei}
\alias{isco88_para_isei}
\title{Indice socioeconomico ISEI a partir do ISCO-88}
\usage{
isco88_para_isei(isco88)
}
\arguments{
\item{isco88}{Vetor de codigos ISCO-88.}
}
\value{
Vetor numerico com o escore ISEI-88.
}
\description{
Indice socioeconomico ISEI a partir do ISCO-88
}
\examples{
isco88_para_isei(c("2221", "5220", "9130"))
}
