% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prestigio.R
\name{isco88_para_prestigio}
\alias{isco88_para_prestigio}
\title{Prestigio de Treiman a partir do ISCO-88}
\usage{
isco88_para_prestigio(isco88)
}
\arguments{
\item{isco88}{Vetor de codigos ISCO-88.}
}
\value{
Vetor numerico.
}
\description{
Prestigio de Treiman a partir do ISCO-88
}
\examples{
isco88_para_prestigio("2221")
}
