% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbo.R
\name{isco88_para_siops}
\alias{isco88_para_siops}
\title{Prestigio ocupacional de Treiman (SIOPS) a partir do ISCO-88}
\usage{
isco88_para_siops(isco88)
}
\arguments{
\item{isco88}{Vetor de codigos ISCO-88.}
}
\value{
Vetor numerico com o escore de prestigio.
}
\description{
Prestigio ocupacional de Treiman (SIOPS) a partir do ISCO-88
}
\examples{
isco88_para_siops(c("2221", "5220", "9130"))
}
