% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isco08.R
\name{tse_para_siops08}
\alias{tse_para_siops08}
\title{SIOPS-08 da ocupação declarada ao TSE}
\usage{
tse_para_siops08(cod)
}
\arguments{
\item{cod}{Vetor de códigos de ocupação do TSE (numérico ou texto).}
}
\value{
Vetor numérico com o escore SIOPS ancorado na ISCO-08.
}
\description{
SIOPS-08 da ocupação declarada ao TSE
}
\examples{
tse_para_siops08(c(111, 169))
}
