% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbo.R
\name{cbo2002_para_siops}
\alias{cbo2002_para_siops}
\title{Prestigio ocupacional SIOPS a partir da CBO-2002}
\usage{
cbo2002_para_siops(cbo)
}
\arguments{
\item{cbo}{Vetor de codigos da CBO-2002, com 6 digitos (ocupacao) ou 4
(familia). Aceita "1111-05", "111105" ou 111105.}
}
\value{
Vetor numerico com o escore SIOPS.
}
\description{
Prestigio ocupacional SIOPS a partir da CBO-2002
}
\examples{
cbo2002_para_siops(c("1111-05", "225120"))
}
