% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{isco08_medidas}
\alias{isco08_medidas}
\title{Medidas ancoradas na ISCO-08}
\format{
\code{data.frame} com as colunas:
\describe{
\item{isco08}{código ISCO-08 de quatro dígitos.}
\item{isei08}{escore ISEI-08, contínuo.}
\item{siops08}{escore SIOPS-08, contínuo.}
}
}
\source{
Módulos \code{isqoisei08.sps} e \code{isqotrei08.sps} do International
Stratification and Mobility File.
\url{http://www.harryganzeboom.nl/ismf/index.htm}
}
\usage{
isco08_medidas
}
\description{
Medidas ancoradas na ISCO-08
}
\examples{
head(isco08_medidas)

# A âncora é outra: um escore da ISCO-08 não é comparável com um da
# ISCO-88, e misturar as duas numa mesma série é o erro mais comum.
nrow(isco08_medidas)
summary(isco08_medidas$isei08)

# Estes escores foram estimados fora do Brasil. Para a régua estimada em
# dado brasileiro, com o mesmo método, veja [isco08_isei_br].
}
\keyword{datasets}
