% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{isco08_medidas}
\alias{isco08_medidas}
\title{Medidas ancoradas na ISCO-08}
\format{
\code{data.frame} com as colunas:
\describe{
\item{isco08}{código ISCO-08 de quatro dígitos.}
\item{isei08}{escore ISEI-08, contínuo.}
\item{siops08}{escore SIOPS-08, contínuo.}
}
}
\source{
Módulos \code{isqoisei08.sps} e \code{isqotrei08.sps} do International
Stratification and Mobility File.
\url{http://www.harryganzeboom.nl/ismf/index.htm}
}
\usage{
isco08_medidas
}
\description{
Medidas ancoradas na ISCO-08
}
\keyword{datasets}
