% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{isco88_isco08}
\alias{isco88_isco08}
\title{Ponte da ISCO-88 para a ISCO-08}
\format{
\code{data.frame} com as colunas:
\describe{
\item{isco88}{código de origem.}
\item{isco08}{código de destino, já truncado como manda a sintaxe.}
\item{n_alternativas}{quantos destinos a Organização Internacional do
Trabalho define para esse código de origem. Vale 1 quando a conversão é
unívoca; acima disso, a tradução é uma escolha entre alternativas, e
convém dizê-lo ao leitor.}
}
}
\source{
Módulo \code{isco8808.sps} do International Stratification and Mobility
File. \url{http://www.harryganzeboom.nl/ismf/index.htm}
}
\usage{
isco88_isco08
}
\description{
Ponte da ISCO-88 para a ISCO-08
}
\keyword{datasets}
