% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prestigio.R
\name{tse_para_prestigio08}
\alias{tse_para_prestigio08}
\title{Prestigio de Treiman ancorado na ISCO-08, a partir do TSE}
\usage{
tse_para_prestigio08(cod)
}
\arguments{
\item{cod}{Vetor de códigos de ocupação do TSE (numérico ou texto).}
}
\value{
Vetor numerico.
}
\description{
Prestigio de Treiman ancorado na ISCO-08, a partir do TSE
}
\examples{
tse_para_prestigio08(c(111, 169))
}
